requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
ffmpeg-python==0.2.0
Flask==3.0.2
Flask-MySQLdb==2.0.0
future==1.0.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
moviepy @ git+https://github.com/Zulko/moviepy.git@bc8d1a831d2d1f61abfdf1779e8df95d523947a5
mysqlclient==2.2.4
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
proglog==0.1.10
psycopg2-binary==2.9.9
requests==2.31.0
SQLAlchemy==2.0.28
sqlalchemy-cockroachdb==2.0.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
